Compiles but no worky

+++ Source/DynaRec/x86/AssemblyWriterX86.cpp
@@ -19,7 +19,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 #include "BuildOptions.h"
 #include "Base/Types.h"
-#include "DyanRec/x86/AssemblyWriterX86.h"
+#include "DynaRec/x86/AssemblyWriterX86.h"
 
 //*****************************************************************************
 //
@@ -71,7 +71,7 @@ void	CAssemblyWriterX86::MUL_EAX_MEM(void * mem)
 {
 	EmitBYTE(0xf7);
 	EmitBYTE(0x25);
-	EmitDWORD((u32)mem);
+	EmitDWORD((uintptr_t)mem);
 }
 
 //*****************************************************************************
@@ -91,7 +91,7 @@ void	CAssemblyWriterX86::ADD_REG_MEM_IDXx4( EIntelReg destreg, void * mem, EInte
 	EmitBYTE(0x03);
 	EmitBYTE(0x04 | (destreg << 3));
 	EmitBYTE(0x85 | (idxreg<<3));
-	EmitDWORD((u32)mem);
+	EmitDWORD((uintptr_t)mem);
 }
 
 //*****************************************************************************
@@ -325,7 +325,7 @@ void CAssemblyWriterX86::CMP_MEM32_I32(const void *p_mem, u32 data)
 {
 	EmitBYTE(0x81);
 	EmitBYTE(0x3d);
-	EmitDWORD(reinterpret_cast< u32 >( p_mem ));
+	EmitDWORD(reinterpret_cast< uintptr_t >( p_mem ));
 	EmitDWORD(data);
 }
 
@@ -336,7 +336,7 @@ void CAssemblyWriterX86::CMP_MEM32_I8(const void *p_mem, u8 data)
 {
 	EmitBYTE(0x83);
 	EmitBYTE(0x3d);
-	EmitDWORD(reinterpret_cast< u32 >( p_mem ));
+	EmitDWORD(reinterpret_cast< uintptr_t >( p_mem ));
 	EmitBYTE(data);
 }
 
@@ -537,7 +537,7 @@ void	CAssemblyWriterX86::CALL_MEM_PLUS_REGx4( void * mem, EIntelReg reg )
 	EmitBYTE(0xFF);
 	EmitBYTE(0x14);
 	EmitBYTE(0x85 | (reg<<3));
-	EmitDWORD((u32)mem);
+	EmitDWORD((uintptr_t)mem);
 }
 
 //*****************************************************************************
@@ -592,7 +592,7 @@ void	CAssemblyWriterX86::MOV_MEM_REG(void * mem, EIntelReg isrc)
 		EmitBYTE(0x89);
 		EmitBYTE((isrc<<3) | 0x05);
 	}
-	EmitDWORD((u32)mem);
+	EmitDWORD((uintptr_t)mem);
 }
 
 //*****************************************************************************
@@ -608,7 +608,7 @@ void	CAssemblyWriterX86::MOV_REG_MEM(EIntelReg reg, const void * mem)
 		EmitBYTE((reg<<3) | 0x05);
 	}
 
-	EmitDWORD((u32)mem);
+	EmitDWORD((uintptr_t)mem);
 }
 
 //*****************************************************************************
@@ -764,7 +764,7 @@ void	CAssemblyWriterX86::MOVI_MEM(void * mem, u32 data)
 {
 	EmitBYTE(0xc7);
 	EmitBYTE(0x05);
-	EmitDWORD((u32)mem);
+	EmitDWORD((uintptr_t)mem);
 	EmitDWORD(data);
 }
 
@@ -775,7 +775,7 @@ void	CAssemblyWriterX86::MOVI_MEM8(void * mem, u8 data)
 {
 	EmitBYTE(0xc6);
 	EmitBYTE(0x05);
-	EmitDWORD((u32)mem);
+	EmitDWORD((uintptr_t)mem);
 	EmitBYTE(data);
 }
 
@@ -822,7 +822,7 @@ void	CAssemblyWriterX86::FCHS()
 void	CAssemblyWriterX86::FILD_MEM( void * pmem )
 {
 	EmitWORD(0x05db);
-	EmitDWORD( u32(pmem) );
+	EmitDWORD(uintptr_t(pmem) );
 }
 
 //*****************************************************************************
@@ -831,7 +831,7 @@ void	CAssemblyWriterX86::FILD_MEM( void * pmem )
 void	CAssemblyWriterX86::FLD_MEMp32( void * pmem )
 {
 	EmitWORD(0x05d9);
-	EmitDWORD( u32(pmem) );
+	EmitDWORD(uintptr_t(pmem) );
 }
 
 //*****************************************************************************
@@ -840,7 +840,7 @@ void	CAssemblyWriterX86::FLD_MEMp32( void * pmem )
 void	CAssemblyWriterX86::FSTP_MEMp32( void * pmem )
 {
 	EmitWORD( 0x1dd9);
-	EmitDWORD( u32(pmem) );
+	EmitDWORD(uintptr_t(pmem) );
 }
 
 //*****************************************************************************
@@ -855,7 +855,7 @@ void	CAssemblyWriterX86::FLD_MEMp64( void * memlo, void * memhi )
 	MOV_MEM_REG(((u8*)&longtemp) + 0, EAX_CODE);
 	MOV_MEM_REG(((u8*)&longtemp) + 4, EDX_CODE);
 	EmitWORD(0x05dd);
-	EmitDWORD( u32(&longtemp) );
+	EmitDWORD(uintptr_t(&longtemp) );
 }
 
 //*****************************************************************************
@@ -866,7 +866,7 @@ void	CAssemblyWriterX86::FSTP_MEMp64( void * memlo, void * memhi )
 	static s64 longtemp;
 
 	EmitWORD(0x1ddd);
-	EmitDWORD( u32(&longtemp) );
+	EmitDWORD(uintptr_t(&longtemp) );
 	MOV_REG_MEM(EAX_CODE, ((u8*)(&longtemp))+0);
 	MOV_REG_MEM(EDX_CODE, ((u8*)(&longtemp))+4);
 	MOV_MEM_REG(((u8*)(memlo)), EAX_CODE);
@@ -879,7 +879,7 @@ void	CAssemblyWriterX86::FSTP_MEMp64( void * memlo, void * memhi )
 void	CAssemblyWriterX86::FISTP_MEMp( void * pmem )
 {
 	EmitWORD( 0x1ddb );
-	EmitDWORD((u32)pmem);
+	EmitDWORD((uintptr_t)pmem);
 }
 
 //*****************************************************************************
